fix: Move tests to proper folders
